[FP and fused units] Renamed units.
