+++ fu/ALUgen_fusedALU_fixedp/test/ALUgenMACFU_test.py
@@ -1,11 +1,10 @@
 """
 ==========================================================================
-FaddRTL_test.py
+_test.py
 ==========================================================================
-Test cases for floating point add unit.
 
-Author : Yanghui Ou
-  Date : Aug 8, 2023
+Author : RJ
+  Date : Jan 6, 2024
 """
 
 from pymtl3                       import *
@@ -16,31 +15,28 @@ from ....lib.test_srcs            import TestSrcRTL
 from ....lib.opt_type             import *
 from ....lib.messages             import *
 from ....mem.const.ConstQueueRTL  import ConstQueueRTL
-from ..FPadd2nd                   import FPadd2nd
-from ...pymtl3_hardfloat.HardFloat.converter_funcs import (floatToFN,
-                                                           fNToFloat)
-
-round_near_even   = 0b000
+from ..ALUgenMACFU                import ALUgenMACFU
 
 def test_elaborate( cmdline_opts ):
   DataType      = mk_data( 16, 1 )
   PredicateType = mk_predicate( 1, 1 )
-  ConfigType    = mk_ctrl()
+  ConfigType    = mk_ctrl(3, 3, 1)
   data_mem_size = 8
-  num_inports   = 2
+  num_inports   = 3
   num_outports  = 1
   FuInType      = mk_bits( clog2( num_inports + 1 ) )
   pickRegister  = [ FuInType( x+1 ) for x in range( num_inports ) ]
   src_in0       = [ DataType(1, 1), DataType(7, 1), DataType(4, 1) ]
   src_in1       = [ DataType(2, 1), DataType(3, 1), DataType(1, 1) ]
+  src_in2       = [ DataType(2, 1), DataType(3, 1), DataType(1, 1) ]
   src_predicate = [ PredicateType(1, 0), PredicateType(1, 0), PredicateType(1, 1) ]
   src_const     = [ DataType(5, 1), DataType(0, 0), DataType(7, 1) ]
   sink_out      = [ DataType(6, 0), DataType(4, 0), DataType(11, 1) ]
   src_opt       = [ ConfigType( OPT_ADD_CONST, b1( 1 ), pickRegister ),
                     ConfigType( OPT_SUB,       b1( 1 ), pickRegister ),
                     ConfigType( OPT_ADD_CONST, b1( 1 ), pickRegister ) ]
-  dut = FPadd2nd( DataType, PredicateType, ConfigType, num_inports,
-                  num_outports, data_mem_size, exp_nbits = 4, sig_nbits = 11 )
+  dut = ALUgenMACFU( DataType, PredicateType, ConfigType, num_inports,
+                     num_outports, data_mem_size )
   dut = config_model_with_cmdline_opts( dut, cmdline_opts, duts=[] )
 
 #-------------------------------------------------------------------------
@@ -51,11 +47,12 @@ class TestHarness( Component ):
 
   def construct( s, FunctionUnit, DataType, PredicateType, ConfigType,
                  num_inports, num_outports, data_mem_size,
-                 src0_msgs, src1_msgs, src_predicate, src_const,
+                 src0_msgs, src1_msgs, src2_msgs, src_predicate, src_const,
                  ctrl_msgs, sink_msgs ):
 
     s.src_in0       = TestSrcRTL( DataType,      src0_msgs     )
     s.src_in1       = TestSrcRTL( DataType,      src1_msgs     )
+    s.src_in2       = TestSrcRTL( DataType,      src2_msgs     )
     s.src_predicate = TestSrcRTL( PredicateType, src_predicate )
     s.src_opt       = TestSrcRTL( ConfigType,    ctrl_msgs     )
     s.sink_out      = TestSinkRTL( DataType,     sink_msgs     )
@@ -69,6 +66,7 @@ class TestHarness( Component ):
 
     connect( s.src_in0.send,       s.dut.recv_in[0]         )
     connect( s.src_in1.send,       s.dut.recv_in[1]         )
+    connect( s.src_in2.send,       s.dut.recv_in[2]         )
     connect( s.src_predicate.send, s.dut.recv_predicate     )
     connect( s.dut.recv_const,     s.const_queue.send_const )
     connect( s.src_opt.send,       s.dut.recv_opt           )
@@ -81,36 +79,34 @@ class TestHarness( Component ):
   def line_trace( s ):
     return s.dut.line_trace()
 
-def mk_float_to_bits_fn( DataType, exp_nbits = 4, sig_nbits = 11 ):
-  return lambda f_value, predicate: (
-      DataType( floatToFN( f_value,
-                           precision = 1 + exp_nbits + sig_nbits ),
-                predicate ) )
-
 def test_add_basic(cmdline_opts):
-  FU            = FPadd2nd
+  FU            = ALUgenMACFU
   exp_nbits     = 4
   sig_nbits     = 11
-  DataType      = mk_data( 1 + exp_nbits + sig_nbits, 1 )
-  f2b           = mk_float_to_bits_fn( DataType, exp_nbits, sig_nbits )
+  DataType      = mk_data( 16, 1 )
   PredicateType = mk_predicate( 1, 1 )
-  ConfigType    = mk_ctrl()
+  ConfigType    = mk_ctrl(3, 3, 1)
   data_mem_size = 8
-  num_inports   = 2
+  num_inports   = 3
   num_outports  = 1
   FuInType      = mk_bits( clog2( num_inports + 1 ) )
   pickRegister  = [ FuInType( x+1 ) for x in range( num_inports ) ]
-  src_in0       = [ f2b(1.1, 1), f2b(7.7, 1), f2b(4.4, 1) ]
-  src_in1       = [ f2b(2.2, 1), f2b(3.3, 1), f2b(1.1, 1) ]
-  src_predicate = [ PredicateType(1, 0), PredicateType(1, 0), PredicateType(1, 1) ]
-  src_const     = [ f2b(5.5, 1), f2b(0, 0), f2b(7.7, 1) ]
-  sink_out      = [ f2b(6.602, 0), f2b(4.4, 0), f2b(12.1, 1) ] # 6.6 -> 6.602
-  src_opt       = [ ConfigType( OPT_FADD_CONST, b1( 1 ), pickRegister ),
-                    ConfigType( OPT_FSUB,       b1( 1 ), pickRegister ),
-                    ConfigType( OPT_FADD_CONST, b1( 1 ), pickRegister ) ]
+  src_in0       = [ DataType(1, 1), DataType(2, 1), DataType(3, 1),DataType(3, 1),DataType(3, 1),DataType(3, 1), DataType(3, 1) ]
+  src_in1       = [ DataType(1, 1), DataType(2, 1), DataType(3, 1),DataType(3, 1),DataType(3, 1),DataType(3, 1), DataType(3, 1) ]
+  src_in2       = [ DataType(1, 1), DataType(2, 1), DataType(3, 1),DataType(3, 1),DataType(3, 1),DataType(3, 1), DataType(3, 1) ]
+  src_predicate = [ PredicateType(1, 1), PredicateType(1, 1), PredicateType(1, 1) ]
+  src_const     = [ DataType(1, 1), DataType(2, 1), DataType(3, 1) ]
+  sink_out      = [ DataType(2, 1), DataType(0, 1), DataType(0, 1),DataType(1, 1),DataType(0, 1),DataType(1, 1), DataType(12, 1) ]
+  src_opt       = [ ConfigType( OPT_ADD, b1( 1 ), pickRegister ),
+                    ConfigType( OPT_SUB, b1( 1 ), pickRegister ),
+                    ConfigType( OPT_LT, b1( 1 ), pickRegister ),
+                    ConfigType( OPT_LTE, b1( 1 ), pickRegister ),
+                    ConfigType( OPT_GT, b1( 1 ), pickRegister ),
+                    ConfigType( OPT_GTE, b1( 1 ), pickRegister ),
+                    ConfigType( OPT_MUL_ADD, b1( 1 ), pickRegister ) ]
   th = TestHarness( FU, DataType, PredicateType, ConfigType,
                     num_inports, num_outports, data_mem_size,
-                    src_in0, src_in1, src_predicate, src_const, src_opt,
+                    src_in0, src_in1, src_in2, src_predicate, src_const, src_opt,
                     sink_out )
   run_sim( th )
